refactor: hoist voice-echo pure helpers to server/lib/voiceEcho.js (#6818)
The tokenizer, trigram builder, and echo thresholds existed as two hand
copies with a "KEEP THE TWO IN SYNC" comment and no guard: server/services/
voice/echo.js (the Whisper/continuous-audio path) and client/src/services/
voiceClient.js (the Web Speech path). Move the pure half (tokenize,
trigramsOf, ECHO_WINDOW_MS, MIN_TOKENS_FOR_ECHO_CHECK, MIN_SHARED_TRIGRAMS)
into a new server/lib/voiceEcho.js leaf; the server echo module now imports
it instead of redefining it, and the client imports the same leaf directly.
The per-socket echo-buffer registry and the client's headset-bypass rule
(looksLikeTtsEcho) are not part of the shared algorithm and stay where they
were.
